update server-identity, use ui-socket
replace direct dom mounting, use ui-socket
convert to typescript
