+++ rundeckapp/grails-spa/packages/ui-trellis/vue.config.app.js
@@ -19,7 +19,7 @@ module.exports = {
     'components/theme':                       { entry: './src/app/components/theme/main.ts'},
     'components/tour':                        { entry: './src/app/components/tour/main.js'},
     'components/version':                     { entry: './src/app/components/version/main.js'},
-    'components/server-identity':             { entry: './src/app/components/server-identity/serverIdentity.js'},
+    'components/server-identity':             { entry: './src/app/components/server-identity/serverIdentity.ts'},
     'pages/storage':                          { entry: './src/app/pages/storage/main.ts'},
     'pages/login':                            { entry: './src/app/pages/login/main.ts'},
     'pages/project-dashboard':                { entry: './src/app/pages/project-dashboard/main.js'},
